pyproject.toml
[project]
name = "qwen-edit-service"
version = "0.2.0"
description = "FastAPI service exposing Qwen-Image-Edit-2511 with dual-GPU bf16 inference"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "xuxiao305" }]

# These are the runtime versions actually deployed; install via deploy/setup_env.sh
# rather than pip from this file (PyTorch wheel selection requires a custom index).
dependencies = [
  "torch==2.5.1",
  "torchvision==0.20.1",
  "diffusers>=0.37.0,<0.40",
  "transformers>=4.49,<5",
  "accelerate>=1.0.0",
  "safetensors>=0.4.5",
  "sentencepiece",
  "protobuf",
  "pillow",
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
]

[project.optional-dependencies]
client = ["requests"]
dev = ["pytest", "ruff"]
